feat: new component products for displaying product
