+++ app/routes/products/route.tsx
@@ -3,6 +3,7 @@ import { useLoaderData } from "@remix-run/react";
 import prisma from "~/_lib/db";
 import ProductCard from "~/components/productCard";
 import { Product } from "../../components/types";
+import Products from "~/components/products";
 
 export const loader: LoaderFunction = async () => {
   const products = await prisma.product.findMany();
@@ -12,14 +13,5 @@ export const loader: LoaderFunction = async () => {
 export default function ProductsPage() {
   const { products } = useLoaderData<{ products: Product[] }>();
 
-  return (
-    <div>
-      <h1>Products</h1>
-      <div className="flex flex-wrap gap-4">
-        {products.map((product) => (
-          <ProductCard key={product.id} product={product} />
-        ))}
-      </div>
-    </div>
-  );
+  return <Products products={products} />;
 }
